+++ forui/lib/src/foundation/rendering.dart
@@ -29,6 +29,27 @@ extension RenderBoxes on RenderBox {
 
 @internal
 extension Alignments on Alignment {
+  static Offset ignoreDirectionalPadding(EdgeInsets padding, Alignment follower, Alignment target) {
+    final Alignment(:x, :y) = follower;
+    // ignore corners that are diagonal.
+    if (x != 0 && y != 0 && x == -target.x && y == -target.y) {
+      return Offset.zero;
+    }
+
+    return Offset(
+      switch (target) {
+        _ when x == -1 && (target.x == x || target.x == 0) => -padding.left,
+        _ when x == 1 && (target.x == x || target.x == 0) => padding.right,
+        _ => 0,
+      },
+        switch (target) {
+          _ when y == -1 && (target.y == y || target.y == 0) => -padding.top,
+          _ when y == 1 && (target.y == y || target.y == 0) => padding.bottom,
+          _ => 0,
+        },
+    );
+  }
+
   Alignment flipX() => switch (this) {
         Alignment.topLeft => Alignment.topRight,
         Alignment.topRight => Alignment.topLeft,